+++ src/igor/igor/igor.h
@@ -151,7 +151,6 @@
 #include <igor/ui/widgets/iWidgetSpacer.h>
 #include <igor/ui/widgets/iWidgetLineTextEdit.h>
 #include <igor/ui/widgets/iWidgetTextEdit.h>
-#include <igor/ui/user_controls/iUserControlAction.h>
 #include <igor/ui/user_controls/iUserControlColorChooser.h>
 #include <igor/ui/user_controls/iUserControlFileChooser.h>