+++ src/tools/mica/src/actions/Actions.h
@@ -29,7 +29,7 @@
 #ifndef __ACTIONS__
 #define __ACTIONS__
 
-#include "ActionContext.h"
+#include "MicaActionContext.h"
 
 /*! delete node action
 */